Convert to meson. Fix compiler warnings etc WIP

+++ libblock/block_lseek.c
@@ -7,7 +7,7 @@
 #include "libblock.h"
 #include "block.h"
 
-int block_lseek(stream_FILE *_file, off_t offset, int whence)
+off_t block_lseek(stream_FILE *_file, off_t offset, int whence)
 {
 	block_FILE *file = (block_FILE*)_file;
 	long new_offset;
